+++ apis/api-core/api-core-solr/src/main/java/org/hoteia/qalingo/core/solr/bean/ProductMarketingSolr.java
@@ -21,7 +21,7 @@ public class ProductMarketingSolr {
 	private Long id;
 
     @Field
-    private Integer scoring;
+    private Long scoring;
     
 	private Boolean isDefault;
 
@@ -87,11 +87,11 @@ public class ProductMarketingSolr {
 		this.id = id;
 	}
 
-    public Integer getScoring() {
+    public Long getScoring() {
         return scoring;
     }
 
-    public void setScoring(Integer scoring) {
+    public void setScoring(Long scoring) {
         this.scoring = scoring;
     }
