+++ apis/api-core/api-core-solr/src/main/java/org/hoteia/qalingo/core/solr/bean/StoreSolr.java
@@ -21,7 +21,7 @@ public class StoreSolr {
 	private Long id;
 	
     @Field
-    private Integer scoring;
+    private Long scoring;
     
     @Field
     private String code;
@@ -70,11 +70,11 @@ public class StoreSolr {
         this.id = id;
     }
 
-    public Integer getScoring() {
+    public Long getScoring() {
         return scoring;
     }
 
-    public void setScoring(Integer scoring) {
+    public void setScoring(Long scoring) {
         this.scoring = scoring;
     }
